requirements.txt
apache-sedona>=1.8.1
pyspark>=3.5.3,<4
overturemaps>=0.19.0
geopandas
matplotlib
# LangChain dependencies  
langgraph
langchain

langchain-openai
# test
pytest>=8.4.0
python-dotenv
pytest-mock

# duckdb
duckdb
boto3